+++ pyklopp/console/commands/init.py
@@ -165,7 +165,7 @@ class InitCommand(Command):
 
 
         self.info('Configuration:')
-        self.info(str(config))
+        self.info(json.dumps(config, indent=2, sort_keys=True))
 
         # Instantiate model
         config['time_model_init_start'] = time.time()
@@ -188,7 +188,7 @@ class InitCommand(Command):
             config_file_path = os.path.join(save_path_base, config_file_name)
             self.info('Writing configuration to "%s"' % config_file_path)
             with open(config_file_path, 'w') as handle:
-                json.dump({'init': config}, handle, indent=2)
+                json.dump({'init': config}, handle, indent=2, sort_keys=True)
 
         self.info('Done.')