+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 # Development Version
 
+## 0.2
+* Logging of evaluation metrics during training
+* Sorting of configuration on standard output and when writing to file
+* New keys in config such as training metrics
+* Started unifying behaviour of loading modules
+
 ## 0.1
 * added commands *init* and *train*
 * custom module loading for model or data set